Holdings no longer persisted to database.
Holdings are only recorded as files for comparison to algorithm which calculates
everything from transactions. The holdings files are stored in a
directory which is symbilicly linked, hence the files are no longer
committed as part of the Git repo. When scraping holdings from the web
site, equivalent files are persisted to store the fund prices, which are
used instead of holdings files when reloading the database.